+++ CHANGELOG.md
@@ -13,10 +13,11 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
-### Added
-- pixReversal description
+
+## [0.13.0] - 2024-09-25
 ### Added
 - installment data to IssuingPurchase resource
+- pixReversal description
 
 ## [0.12.0] - 2024-07-01
 ### Added